Create iterate through array with loops
We have defined a function, filteredArray, which takes arr, a nested array, and elem as arguments, and returns a new array. elem represents an element that may or may not be present on one or more of the arrays nested within arr.Modify the function, using a for loop, to return a filtered version of the passed array such that any array nested within arr containing elem has been removed.